move project keywords to a sparate component
