+++ client/src/features/ProjectPageV2/ProjectPageContent/Settings/ProjectSettings.tsx
@@ -18,7 +18,7 @@
 
 import cx from "classnames";
 import { useCallback, useContext, useEffect, useState } from "react";
-import { Diagram3Fill, Pencil, PlusLg, Sliders } from "react-bootstrap-icons";
+import { Diagram3Fill, Pencil, Sliders } from "react-bootstrap-icons";
 import { Controller, useForm } from "react-hook-form";
 import { generatePath, useLocation, useNavigate } from "react-router";
 import {
@@ -54,10 +54,9 @@ import type {
 import useProjectPermissions from "../../utils/useProjectPermissions.hook";
 import ProjectSessionSecrets from "../SessionSecrets/ProjectSessionSecrets";
 import ProjectPageDelete from "./ProjectDelete";
+import ProjectKeywordsFormField from "./ProjectKeywordsFormField";
 import ProjectPageSettingsMembers from "./ProjectSettingsMembers";
 import ProjectUnlinkTemplate from "./ProjectUnlinkTemplate";
-import KeywordContainer from "~/components/keywords/KeywordContainer";
-import KeywordBadge from "~/components/keywords/KeywordBadge";
 
 function notificationProjectUpdated(
   notifications: NotificationsManager,
@@ -401,99 +400,12 @@ function ProjectSettingsForm({ project }: ProjectPageSettingsProps) {
         <PermissionsGuard
           disabled={null}
           enabled={
-            <div>
-              <Label className="form-label" for="project-keywords">
-                Keywords
-              </Label>
-              <div className={cx("input-group", "input-group-sm", "mb-2")}>
-                <Controller
-                  name="keyword"
-                  control={control}
-                  render={({ field }) => (
-                    <>
-                      <input
-                        id="keyword"
-                        placeholder="Add new keyword"
-                        type="string"
-                        {...field}
-                        className={cx(
-                          "form-control",
-                          errors.keyword && "is-invalid"
-                        )}
-                        data-cy="project-settings-keyword-input"
-                        onChange={(e) => {
-                          field.onChange(e);
-                        }}
-                        onKeyDown={(e) => {
-                          if (e.key === "Enter" && field.value) {
-                            e.preventDefault();
-                            const newValue = field.value.trim();
-                            const currentKeywords = getValues("keywords");
-                            if (!currentKeywords.includes(newValue)) {
-                              const newKeywords = [
-                                ...currentKeywords,
-                                newValue,
-                              ];
-                              setValue("keywords", newKeywords);
-                            }
-                            setValue("keyword", "");
-                          }
-                        }}
-                      />
-                      <Button
-                        color={field.value ? "primary" : "outline-primary"}
-                        disabled={!field.value}
-                        data-cy="project-settings-keyword-button"
-                        onClick={() => {
-                          if (field.value) {
-                            const newValue = field.value.trim();
-                            const currentKeywords = getValues("keywords");
-                            if (!currentKeywords.includes(newValue)) {
-                              const newKeywords = [
-                                ...currentKeywords,
-                                newValue,
-                              ];
-                              setValue("keywords", newKeywords);
-                            }
-                            setValue("keyword", "");
-                          }
-                        }}
-                        type="button"
-                      >
-                        <PlusLg className={cx("bi", "me-1")} />
-                        Add
-                      </Button>
-                    </>
-                  )}
-                />
-              </div>
-              <Controller
-                name="keywords"
-                control={control}
-                render={({ field }) => (
-                  <>
-                    {field.value && field.value.length > 0 && (
-                      <KeywordContainer data-cy="project-settings-keywords">
-                        {getValues("keywords").map((keyword, index) => (
-                          <KeywordBadge
-                            data-cy="project-settings-keyword"
-                            key={index}
-                            removeHandler={() => {
-                              const newKeywords = getValues("keywords").filter(
-                                (k) => k !== keyword
-                              );
-                              setValue("keywords", newKeywords);
-                            }}
-                          >
-                            {keyword}
-                          </KeywordBadge>
-                        ))}
-                      </KeywordContainer>
-                    )}
-                  </>
-                )}
-              />
-            </div>
+            <ProjectKeywordsFormField
+              control={control}
+              errors={errors}
+              getValues={getValues}
+              setValue={setValue}
+            />
           }
           requestedPermission="write"
           userPermissions={permissions}
